requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
alabaster==0.7.16
    # via sphinx
amqp==5.2.0
    # via kombu
anyio==4.4.0
    # via
    #   starlette
    #   watchfiles
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.8.1
    # via
    #   django
    #   django-stubs
asttokens==2.4.1
    # via stack-data
async-timeout==4.0.3
    # via redis
babel==2.16.0
    # via sphinx
billiard==4.2.0
    # via celery
celery==5.4.0
    # via
    #   celery-redbeat
    #   django-celery-beat
    #   flower
celery-redbeat==2.2.0
certifi==2024.7.4
    # via requests
cffi==1.17.0
    # via
    #   argon2-cffi-bindings
    #   cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   djlint
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   djlint
    #   sphinx-autobuild
coverage==7.6.1
    # via django-coverage-plugin
crispy-bootstrap5==2024.2
cron-descriptor==1.4.3
    # via django-celery-beat
cryptography==43.0.0
    # via fido2
cssbeautifier==1.15.1
    # via djlint
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
diff-match-patch==20230430
    # via django-import-export
distlib==0.3.8
    # via virtualenv
django==5.0.8
    # via
    #   crispy-bootstrap5
    #   django-allauth
    #   django-anymail
    #   django-celery-beat
    #   django-ckeditor
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-import-export
    #   django-jazzmin
    #   django-js-asset
    #   django-model-utils
    #   django-redis
    #   django-stubs
    #   django-stubs-ext
    #   django-timezone-field
django-allauth==64.0.0
django-anymail==11.1
django-celery-beat==2.6.0
django-ckeditor==6.7.1
django-coverage-plugin==3.1.0
django-crispy-forms==2.3
    # via crispy-bootstrap5
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-import-export==4.1.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
    # via django-ckeditor
django-model-utils==4.5.1
django-redis==5.4.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
    # via django-stubs
django-timezone-field==7.0
    # via django-celery-beat
djlint==1.34.1
docutils==0.21.2
    # via sphinx
editorconfig==0.12.4
    # via
    #   cssbeautifier
    #   jsbeautifier
exceptiongroup==1.2.2
    # via
    #   anyio
    #   ipython
    #   pytest
executing==2.0.1
    # via stack-data
factory-boy==3.3.0
faker==26.3.0
    # via factory-boy
fido2==1.1.3
filelock==3.15.4
    # via virtualenv
flower==2.0.1
gunicorn==22.0.0
    # via uvicorn-worker
h11==0.14.0
    # via uvicorn
hiredis==3.0.0
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
httptools==0.6.1
humanize==4.10.0
    # via flower
identify==2.6.0
    # via pre-commit
idna==3.7
    # via
    #   anyio
    #   requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
ipython==8.26.0
    # via ipdb
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via sphinx
jsbeautifier==1.15.1
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.25
    # via djlint
kombu==5.4.0
    # via celery
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.7
    # via ipython
mypy==1.11.1
mypy-extensions==1.0.0
    # via mypy
nodeenv==1.9.1
    # via pre-commit
packaging==24.1
    # via
    #   gunicorn
    #   pytest
    #   pytest-sugar
    #   sphinx
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via djlint
pexpect==4.9.0
    # via ipython
pillow==10.4.0
pip==24.1.2
platformdirs==4.2.2
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.8.0
prometheus-client==0.20.0
    # via flower
prompt-toolkit==3.0.47
    # via
    #   click-repl
    #   ipython
psycopg==3.2.1
psycopg-c==3.2.1
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via
    #   ipython
    #   sphinx
pypng==0.20220715.0
    # via qrcode
pytest==8.3.2
    # via
    #   pytest-django
    #   pytest-sugar
pytest-django==4.8.0
pytest-sugar==1.0.0
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   celery-redbeat
    #   faker
    #   python-crontab
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
    # via flower
pyyaml==6.0.2
    # via
    #   djlint
    #   pre-commit
qrcode==7.4.2
redis==5.0.8
    # via
    #   celery-redbeat
    #   django-redis
regex==2023.12.25
    # via djlint
requests==2.32.3
    # via
    #   django-anymail
    #   sphinx
ruff==0.5.7
setuptools==70.3.0
six==1.16.0
    # via
    #   asttokens
    #   cssbeautifier
    #   jsbeautifier
    #   python-dateutil
sniffio==1.3.1
    # via anyio
snowballstemmer==2.2.0
    # via sphinx
sphinx==7.4.7
    # via sphinx-autobuild
sphinx-autobuild==2024.4.16
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
sqlparse==0.5.1
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
starlette==0.38.2
    # via sphinx-autobuild
tablib==3.5.0
    # via django-import-export
tenacity==9.0.0
    # via celery-redbeat
termcolor==2.4.0
    # via pytest-sugar
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   django-stubs
    #   djlint
    #   ipdb
    #   mypy
    #   pytest
    #   sphinx
tornado==6.4.1
    # via flower
tqdm==4.66.5
    # via djlint
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
types-pyyaml==6.0.12.20240808
    # via django-stubs
typing-extensions==4.12.2
    # via
    #   anyio
    #   asgiref
    #   django-stubs
    #   django-stubs-ext
    #   ipython
    #   mypy
    #   psycopg
    #   qrcode
    #   uvicorn
tzdata==2024.1
    # via
    #   celery
    #   django-celery-beat
urllib3==2.2.2
    # via
    #   django-anymail
    #   requests
uvicorn==0.30.5
    # via
    #   sphinx-autobuild
    #   uvicorn-worker
uvicorn-worker==0.2.0
uvloop==0.19.0
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.26.3
    # via pre-commit
watchdog==4.0.2
watchfiles==0.21.0
    # via sphinx-autobuild
wcwidth==0.2.13
    # via prompt-toolkit
websockets==12.0
    # via sphinx-autobuild
werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.7.0
